Reorg notifications into domains/ (full structure, template)
First domain moved to the target structure, establishing the pattern for the
rest:
- domains/notifications/tables.ts  — the table def, composed into schema.ts
- domains/notifications/model.ts    — plain helpers/constants + the type
- domains/notifications/queries.ts  — read handlers (plain, take QueryCtx)
- domains/notifications/mutations.ts — write handlers + bulk continuations
- notifications.ts (root)           — thin facade: Convex query/mutation/
  internalMutation wrappers at the SAME export names, plus a re-export of
  clearBookingNotifications for booking.

Behavior-preserving: api.d.ts diff is purely additive (the four new module-map
entries); every api.notifications.* / internal.notifications.* path and arg
validator is unchanged, so scheduled continuations and clients keep resolving.
Backend 100 unit + 25 integration + types green.
