+++ packages/backend/convex/_generated/api.d.ts
@@ -18,6 +18,10 @@ import type * as calendar from "../calendar.js";
 import type * as calendarSync from "../calendarSync.js";
 import type * as contacts from "../contacts.js";
 import type * as crons from "../crons.js";
+import type * as domains_notifications_model from "../domains/notifications/model.js";
+import type * as domains_notifications_mutations from "../domains/notifications/mutations.js";
+import type * as domains_notifications_queries from "../domains/notifications/queries.js";
+import type * as domains_notifications_tables from "../domains/notifications/tables.js";
 import type * as googleSync from "../googleSync.js";
 import type * as healthCheck from "../healthCheck.js";
 import type * as http from "../http.js";
@@ -60,6 +64,10 @@ declare const fullApi: ApiFromModules<{
   calendarSync: typeof calendarSync;
   contacts: typeof contacts;
   crons: typeof crons;
+  "domains/notifications/model": typeof domains_notifications_model;
+  "domains/notifications/mutations": typeof domains_notifications_mutations;
+  "domains/notifications/queries": typeof domains_notifications_queries;
+  "domains/notifications/tables": typeof domains_notifications_tables;
   googleSync: typeof googleSync;
   healthCheck: typeof healthCheck;
   http: typeof http;
