+++ packages/backend/convex/schema.ts
@@ -1,6 +1,8 @@
 import { defineSchema, defineTable } from "convex/server";
 import { v } from "convex/values";
 
+import { notificationTables } from "./domains/notifications/tables";
+
 /** A guest on an event. Mirrors the subset of Google's attendee object we keep;
  * `responseStatus` is "needsAction" | "declined" | "tentative" | "accepted".
  * Shared by the `events` table and the mutation validators that write to it. */
@@ -383,27 +385,8 @@ export default defineSchema({
     .index("by_status_and_end", ["status", "endMs"])
     .index("by_token", ["token"]),
 
-  // One row per in-app notification for a host. Written by the events that a
-  // host should hear about (currently a new booking request) and read by the
-  // header notification bell. Dismissing a notification hard-deletes its row, so
-  // the feed and this table never drift apart.
-  notifications: defineTable({
-    // Recipient. Matches `bookings.hostUserId` / the auth user's `_id`.
-    userId: v.string(),
-    type: v.union(v.literal("booking_requested")),
-    title: v.string(),
-    body: v.optional(v.string()),
-    // The booking this notification points at, so a click can open its panel.
-    bookingId: v.optional(v.id("bookings")),
-    read: v.boolean(),
-    createdAt: v.number(),
-  })
-    .index("by_user_and_created", ["userId", "createdAt"])
-    .index("by_user_and_read", ["userId", "read"])
-    .index("by_user_and_read_and_created", ["userId", "read", "createdAt"])
-    // Lets a booking's lifecycle (accept / decline / expire) clear the request
-    // notification it spawned.
-    .index("by_booking", ["bookingId"]),
+  // Notifications domain tables (see domains/notifications/tables.ts).
+  ...notificationTables,
 
   // Fixed-window counters guarding the one mutation anonymous callers can
   // reach. Keyed by requester email and by page slug — a Convex mutation has no
